Move paginations tests into the list builder test.
