+++ domain/tests/src/Functional/DomainListBuilderTest.php
@@ -24,8 +24,8 @@ class DomainListBuilderTest extends DomainTestBase {
   protected function setUp() {
     parent::setUp();
 
-    // Create 5 domains.
-    $this->domainCreateTestDomains(5);
+    // Create 150 domains.
+    $this->domainCreateTestDomains(150);
   }
 
   /**
@@ -45,10 +45,22 @@ class DomainListBuilderTest extends DomainTestBase {
     $this->assertSession()->statusCodeEquals(200);
 
     // Check that links are printed.
-    foreach ($this->getDomains() as $domain) {
+    foreach ($this->getPaginatedDomains() as $domain) {
       $href = 'admin/config/domain/edit/' . $domain->id();
-      $this->assertSession()->linkByHrefExists($href, 0, 'Link found');
+      $this->assertSession()->linkByHrefExists($href, 0, 'Link found ' . $href);
       $this->assertSession()->assertEscaped($domain->label());
+      // Check for pagination.
+      $this->checkPagination();
+    }
+
+    // Go to page 2.
+    $this->clickLink('Next');
+    foreach ($this->getPaginatedDomains(1) as $domain) {
+      $href = 'admin/config/domain/edit/' . $domain->id();
+      $this->assertSession()->linkByHrefExists($href, 0, 'Link found ' . $href);
+      $this->assertSession()->assertEscaped($domain->label());
+      // Check for pagination.
+      $this->checkPagination();
     }
 
     // Now login as a user with limited rights.
@@ -72,9 +84,9 @@ class DomainListBuilderTest extends DomainTestBase {
     $this->assertSession()->statusCodeEquals(200);
 
     // Check that links are printed.
-    $path = 'admin/config/domain/';
+    $path = 'admin/config/domain';
     $this->drupalGet($path);
-    foreach ($this->getDomains() as $domain) {
+    foreach ($this->getPaginatedDomains() as $domain) {
       $href = 'admin/config/domain/edit/' . $domain->id();
       if (in_array($domain->id(), $ids)) {
         $this->assertSession()->linkByHrefExists($href, 0, 'Link found');
@@ -84,10 +96,12 @@ class DomainListBuilderTest extends DomainTestBase {
         $this->assertSession()->linkByHrefNotExists($href, 'Link not found');
         $this->assertSession()->assertEscaped($domain->label());
       }
+      // Check for pagination.
+      $this->checkPagination();
     }
 
     // Check access to the pages/routes.
-    foreach ($this->getDomains() as $domain) {
+    foreach ($this->getPaginatedDomains() as $domain) {
       $path = 'admin/config/domain/edit/' . $domain->id();
       $this->drupalGet($path);
       if (in_array($domain->id(), $ids)) {
@@ -96,6 +110,23 @@ class DomainListBuilderTest extends DomainTestBase {
       else {
         $this->assertSession()->statusCodeEquals(403);
       }
+    }
+
+    // Go to page 2.
+    $this->drupalGet('admin/config/domain');
+    $this->clickLink('Next');
+    foreach ($this->getPaginatedDomains(1) as $domain) {
+      $href = 'admin/config/domain/edit/' . $domain->id();
+      if (in_array($domain->id(), $ids)) {
+        $this->assertSession()->linkByHrefExists($href, 0, 'Link found');
+        $this->assertSession()->assertEscaped($domain->label());
+      }
+      else {
+        $this->assertSession()->linkByHrefNotExists($href, 'Link not found');
+        $this->assertSession()->assertEscaped($domain->label());
+      }
+      // Check for pagination.
+      $this->checkPagination();
     }
 
     // Now login as a user with more limited rights.
@@ -119,9 +150,9 @@ class DomainListBuilderTest extends DomainTestBase {
     $this->assertSession()->statusCodeEquals(200);
 
     // Check that domains are listed and links are printed.
-    $path = 'admin/config/domain/';
+    $path = 'admin/config/domain';
     $this->drupalGet($path);
-    foreach ($this->getDomains() as $domain) {
+    foreach ($this->getPaginatedDomains() as $domain) {
       $href = 'admin/config/domain/edit/' . $domain->id();
       if (in_array($domain->id(), $ids)) {
         $this->assertSession()->linkByHrefExists($href, 0, 'Link found');
@@ -131,10 +162,12 @@ class DomainListBuilderTest extends DomainTestBase {
         $this->assertSession()->linkByHrefNotExists($href, 'Link not found');
         $this->assertSession()->assertNoEscaped($domain->label());
       }
+      // Check for pagination.
+      $this->checkNoPagination();
     }
 
     // Check access to the pages/routes.
-    foreach ($this->getDomains() as $domain) {
+    foreach ($this->getPaginatedDomains() as $domain) {
       $path = 'admin/config/domain/edit/' . $domain->id();
       $this->drupalGet($path);
       if (in_array($domain->id(), $ids)) {
@@ -143,6 +176,37 @@ class DomainListBuilderTest extends DomainTestBase {
       else {
         $this->assertSession()->statusCodeEquals(403);
       }
+    }
+
+  }
+
+  /**
+   * Returns an array of domains, paginated and sorted by weight.
+   *
+   * @param int $page
+   *   The page number to return.
+   */
+  private function getPaginatedDomains($page = 0) {
+    $limit = 50;
+    $offset = $page * $limit;
+    return array_slice($this->getDomainsSorted(), $offset, $limit);
+  }
+
+  /**
+   * Checks that pagination links appear, as expected.
+   */
+  private function checkPagination() {
+    foreach (['?page=0', '?page=1', '?page=2'] as $href) {
+      $this->assertSession()->linkByHrefExists($href, 0, 'Link found');
+    }
+  }
+
+  /**
+   * Checks that pagination links do not appear, as expected.
+   */
+  private function checkNoPagination() {
+    foreach (['?page=0', '?page=1', '?page=2'] as $href) {
+      $this->assertSession()->linkByHrefNotExists($href, 0, 'Link not found');
     }
   }
